requirements.txt
# Ford Intelligence OS - Dependencies
# Stack decisions from /plan-eng-review 2026-03-28

# Core
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Module 1 - Competitive Radar (Scraper)
# playwright — install locally only: pip install playwright && playwright install chromium

# Module 2 - Retention Engine
pandas>=2.0.0
scikit-learn>=1.3.0

# NL Query - SQL Generation
openai>=1.0.0

# Dashboard
streamlit>=1.30.0

# Utilities
rich>=13.0.0
